+++ whale_streaming/client/src/components/main/MainCenter.vue
@@ -4,7 +4,16 @@
         <div class="mainContentFrame">
             <div class="mainContent">
                 <div class="mainContentMargin">
-                    <router-view :key="$route.fullPath" :libraries="libraries" @update-library="getUserLibraries"></router-view>
+                    <!-- Router View에 라이브러리 정보 전달 -->
+                    <router-view
+                        :key="$route.fullPath"
+                        :libraries="libraries"
+                        :change-background="changeBackground"
+                        :play-player="playPlayer"
+                        :scroll-content="scrollContent"
+                        :check-scroll="checkScroll"
+                        @update-library="getUserLibraries"
+                    />
                 </div>
             </div>
         </div>
@@ -29,9 +38,10 @@ export default {
         this.getUserLibraries();
     },
     methods: {
+        // 라이브러리 정보를 받아 오는 함수
         async getUserLibraries() {
             const result = await fetch('/whale/streaming/getLibraries');
-            if (await result.ok) {
+            if (result.ok) {
                 const data = await result.json();
                 if (data.items && data.items.length > 0) {
                     this.libraries = data.items;
@@ -40,6 +50,62 @@ export default {
                 }
             } else {
                 console.error('Failed to fetch user top items:', result.statusText);
+            }
+        },
+        // 랜덤 RGB 색상 생성 함수
+        getRandomColor() {
+            const r = Math.floor(Math.random() * 256);
+            const g = Math.floor(Math.random() * 256);
+            const b = Math.floor(Math.random() * 256);
+            console.log("Generated RGB values:", r, g, b);
+            return `rgb(${r}, ${g}, ${b})`;
+        },
+        // 배경색 변화 함수
+        changeBackground() {
+            if (localStorage.getItem('darkmodeOn') === "1") {document.querySelector('.mainContent').style.backgroundImage = `linear-gradient(${this.getRandomColor()} 0%, rgb(17, 18, 17) 100%)`;}
+            else {document.querySelector('.mainContent').style.backgroundImage = `linear-gradient(${this.getRandomColor()} 0%, rgb(249, 250, 249) 100%)`;}
+        },
+        // 재생 함수
+        async playPlayer(i) {
+            await fetch(`/whale/streaming/play?uri=${ i }&device_id=${ sessionStorage.device_id }`);
+        },
+        // 스크롤 버튼을 위한 함수
+        updateScrollButtons(containerSelector) {
+            const container = document.querySelector(containerSelector);
+            if (!container) return;
+
+            const scrollLeftBtn = container.parentNode.querySelector('.artistDetailSlideButton.left');
+            const scrollRightBtn = container.parentNode.querySelector('.artistDetailSlideButton.right');
+
+            if (container.scrollLeft > 0) {
+                scrollLeftBtn.classList.remove('hidden');
+            } else {
+                scrollLeftBtn.classList.add('hidden');
+            }
+
+            const maxScrollLeft = container.scrollWidth - container.clientWidth;
+            if (container.scrollLeft < maxScrollLeft) {
+                scrollRightBtn.classList.remove('hidden');
+            } else {
+                scrollRightBtn.classList.add('hidden');
+            }
+        },
+        scrollContent(containerSelector, direction) {
+            const container = document.querySelector(containerSelector);
+            if (!container) return;
+            
+            const scrollAmount = direction === 'left' ? -210 : 210;
+            container.scrollBy({ left: scrollAmount, behavior: 'smooth' });
+
+            setTimeout(() => {
+                this.updateScrollButtons(containerSelector);
+            }, 300);
+        },
+        checkScroll(containerSelector) {
+            const container = document.querySelector(containerSelector);
+            if (container) {
+                this.updateScrollButtons(containerSelector);
+                container.addEventListener('scroll', () => this.updateScrollButtons(containerSelector));
             }
         },
     },
