+++ whale_streaming/client/src/components/main/MainDetailAlbum.vue
@@ -82,6 +82,9 @@
 
 <script>
 export default {
+    props: {
+        changeBackground: {type: Function, default() {return 'Default function'}},
+    },
     data() {
         return {
             artist: null,
@@ -132,18 +135,6 @@ export default {
                 });
             }
             this.isPlayed[j] = !this.isPlayed[j];
-        },
-        getRandomColor() {
-            // 랜덤 RGB 색상 생성 함수
-            const r = Math.floor(Math.random() * 256);
-            const g = Math.floor(Math.random() * 256);
-            const b = Math.floor(Math.random() * 256);
-            console.log("Generated RGB values:", r, g, b);
-            return `rgb(${r}, ${g}, ${b})`;
-        },
-        changeBackground() {
-            if (localStorage.getItem('darkmodeOn') === "1") {document.querySelector('.mainContent').style.backgroundImage = `linear-gradient(${this.getRandomColor()} 0%, rgb(17, 18, 17) 100%)`;}
-            else {document.querySelector('.mainContent').style.backgroundImage = `linear-gradient(${this.getRandomColor()} 0%, rgb(249, 250, 249) 100%)`;}
         },
         addIsShow(i) {
             this.isPlayed.push(false);
