+++ whale_streaming/client/src/components/main/MainPlaylist.vue
@@ -145,6 +145,7 @@
 export default {
     props: {
         libraries: Object,
+        changeBackground: {type: Function, default() {return 'Default function'}},
     },
     data() {
         return {
@@ -233,18 +234,6 @@ export default {
                 }
                 this.isPlayed[j] = !this.isPlayed[j];
             }
-        },
-        getRandomColor() {
-            // 랜덤 RGB 색상 생성 함수
-            const r = Math.floor(Math.random() * 256);
-            const g = Math.floor(Math.random() * 256);
-            const b = Math.floor(Math.random() * 256);
-            console.log("Generated RGB values:", r, g, b);
-            return `rgb(${r}, ${g}, ${b})`;
-        },
-        changeBackground() {
-            if (localStorage.getItem('darkmodeOn') === "1") {document.querySelector('.mainContent').style.backgroundImage = `linear-gradient(${this.getRandomColor()} 0%, rgb(17, 18, 17) 100%)`;}
-            else {document.querySelector('.mainContent').style.backgroundImage = `linear-gradient(${this.getRandomColor()} 0%, rgb(249, 250, 249) 100%)`;}
         },
         addIsShow(i) {
             this.isPlayed.push(false);
