+++ whale_streaming/server/models/spotify/dao/fetchSpotify.js
@@ -1,3 +1,4 @@
+// Spotify API에 정보를 요청 및 응답하는 함수
 async function fetchWebApi(req, endpoint, method, body) {
     const response = await fetch(`https://api.spotify.com/${endpoint}`, {
         headers: {
