+++ whale_streaming/server/models/spotify/info/getAlbumInfo.js
@@ -1,3 +1,4 @@
+// 앨범 정보를 요청하는 함수
 const fetchSpotify = require('../dao/fetchSpotify');
 
 async function getAlbumInfo(req,res) {
