+++ whale_streaming/server/models/spotify/info/getArtistInfo.js
@@ -1,3 +1,4 @@
+// 아티스트 정보를 요청하는 함수
 const fetchSpotify = require('../dao/fetchSpotify');
 
 async function getArtistInfo(req,res) {
