+++ whale_streaming/server/models/spotify/users/getPlaylist.js
@@ -1,3 +1,4 @@
+// 플레이리스트 정보를 요청하는 함수
 const fetchSpotify = require('../dao/fetchSpotify');
 
 async function fetchWebApi(req, endpoint, method, body) {
