+++ whale_streaming/server/models/spotify/users/getUserLibraries.js
@@ -1,3 +1,4 @@
+// 유저의 라이브러리 정보를 요청하는 함수
 const fetchSpotify = require('../dao/fetchSpotify');
 
 async function getUserLibraries(req,res) {
