+++ whale_streaming/server/service/spotify/libraryService.js
@@ -1,3 +1,4 @@
+// 라이브러리 정보를 요청하는 서비스
 const userLibraries = require('../../models/spotify/users/getUserLibraries');
 
 async function libraryService(req,res) {
