+++ whale_streaming/server/service/spotify/searchService.js
@@ -1,3 +1,4 @@
+// 검색 정보를 요청하는 서비스
 const search = require('../../models/spotify/search/search');
 
 async function searchService(req,res) {
